requirements.txt
web3
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
git+https://github.com/ZeroEkkusu/foundrycli.py.git
